changed find my friends to work with phone number
If you invite a friend via phone number, the phone number (in an international format) is saved within the "invitationFromHandles" icloud data. 

So I changed the find my friends to work with a phone number. 
Previously everything without an "@" was filtered out, now it needs to contain an "@" or start with an "+".